Reorganize source
